Move device helpers to utils
