+++ tts/chatterbox.py
@@ -2,35 +2,11 @@ import torch
 import torchaudio as ta
 from typing import List
 from chatterbox.tts import ChatterboxTTS
+from utils import get_device  # re-exported for convenience
 
 MAX_CHARS = 300
 
 
-def get_device() -> str:
-    """Detect and return the best available device for torch: 'cuda', then 'mps', else 'cpu'."""
-    if torch.cuda.is_available():
-        return "cuda"
-    elif torch.backends.mps.is_available():
-        return "mps"
-    else:
-        return "cpu"
-
-
-def get_map_location():
-    """Returns the torch.device object for loading models on the detected device."""
-    return torch.device(get_device())
-
-# Expose and patch torch.load to always use the correct map_location unless overridden
-# Usage: torch_load_original(..., map_location=get_map_location())
-torch_load_original = torch.load
-
-def patched_torch_load(*args, **kwargs):
-    """Patched torch.load that uses the correct map_location unless overridden."""
-    if 'map_location' not in kwargs:
-        kwargs['map_location'] = get_map_location()
-    return torch_load_original(*args, **kwargs)
-
-torch.load = patched_torch_load
 
 def synthesize_long_text(text: str, model: ChatterboxTTS, max_chars: int = MAX_CHARS) -> torch.Tensor:
     """Generate audio for text longer than Chatterbox's limit."""
@@ -60,3 +36,6 @@ def generate(lines: List[str], output_file: str, model: ChatterboxTTS, max_chars
     waveform = synthesize_long_text(text, model, max_chars)
     ta.save(output_file, waveform, model.sr)
     return output_file
+
+
+__all__ = ["generate", "synthesize_long_text", "MAX_CHARS", "get_device"]
